Add Filter page, where you ca take books order by date, ratungs and other. Update Profile logic: Implement ability to change the profile picture, etc.

+++ BookStore.BAL/Services/FileService.cs
@@ -7,7 +7,7 @@ public static class FileService
     public static async Task<string> Save(IFormFile file)
     {
         bool isImage = file.ContentType.StartsWith("image");
-        string type = isImage ? "images" : "files";
+        string type = isImage ? "images/book" : "files";
         string filePath = Path.Combine($"/{type}/", file.FileName);
         using (var fileStream = new FileStream("wwwroot" + filePath, FileMode.Create))
         {
@@ -17,7 +17,18 @@ public static class FileService
     }
     public static void Delete(string sourceFile)
     {
+        if (sourceFile.Contains("/server/")) return;
         if (File.Exists("wwwroot" + sourceFile))
             File.Delete("wwwroot" + sourceFile);
+    }
+
+    public static async Task<string> SaveProfilePic(IFormFile img)
+    {
+        string filePath = Path.Combine($"/images/profile/", img.FileName);
+        using (var fileStream = new FileStream("wwwroot" + filePath, FileMode.Create))
+        {
+            await img.CopyToAsync(fileStream);
+        }
+        return filePath;
     }
 }
